Add useChart
