+++ packages/@orchidui-vue/src/Charts/LineChart/OcLineChart.vue
@@ -4,7 +4,8 @@
 
 <script setup lang="ts">
 import * as echarts from "echarts";
-import { computed, onMounted, onUnmounted, ref, watch } from "vue";
+import { computed, onMounted, ref } from "vue";
+import { useChart } from "@/orchidui/composables/useChart.js";
 
 const props = defineProps({
   showTooltip: Boolean,
@@ -135,54 +136,14 @@ const options = computed(() => ({
     },
   ],
 }));
-const myChart = ref();
 const lineChart = ref();
 
-const toggleLegendName = (name) => {
-  myChart.value.dispatchAction({
-    type: "legendToggleSelect",
-    name,
-  });
-};
-
-const renderChart = () => {
-  myChart.value = echarts.init(lineChart.value);
-  myChart.value.setOption(options.value);
-};
-
-const resizeChart = () => {
-  if (myChart.value) {
-    myChart.value.dispose();
-  }
-  renderChart();
-};
-
-defineExpose({
-  toggleLegendName,
-});
+const { chart } = useChart(lineChart, options);
 
 onMounted(() => {
-  renderChart();
-
-  myChart.value.getZr().on("globalout", () => {
+  chart.value.getZr().on("globalout", () => {
     markLineData.value.index = 0;
     markLineData.value.value = 0;
   });
-
-  window.addEventListener("resize", resizeChart);
 });
-onUnmounted(() => {
-  if (myChart.value) {
-    myChart.value.dispose();
-  }
-
-  window.removeEventListener("resize", resizeChart);
-});
-watch(
-  () => options.value,
-  (val) => {
-    myChart.value.setOption(val);
-  },
-  { deep: true },
-);
 </script>
